[IMP] account_edi: improve how we handle attachments for the email
When assembling the email we are leaving the assignment of the attachments
to another method that we can inherit and then modify instead of doing it
in the same method to attach them.
The idea is that in any localization we need only to modify that method
in order to add new attachments to the mail template.

closes odoo/odoo#80484

X-original-commit: d47ba1563bc74cec6d4d22ab9824c1e67758fbdc
Related: odoo/enterprise#22564
Signed-off-by: Laurent Smet <las@openerp.com>
Signed-off-by: Josse Colpaert <jco@openerp.com>