[IMP] project: get the display_project_id if parent is subtask
Before this commit, when the user creates a subtask in subtask then we
take the project_id but if the user has changed the display_project_id,
the subtask of this task has not the correct project.

This commit gives the project_id by default if the parent task is not a
subtask or the display_project_id of this parent task is not defined
otherwise it gives the display_project_id.

task-2232042

closes #51471