[FIX] web_editor: ensure editor in iframe can be scrolled on crop image
The overlay around an image when cropping it is positioned absolutely.
If it has no relative ancestor, it will cover the entire body. This is
what happend in mass_mailing, making it impossible to scroll the page
while cropping. This commit addresses the issue by positioning that
editable relatively so that the overlay never goes beyond its bounds.

closes odoo/odoo#69878

X-original-commit: 85e2ae070bb6ae7d2c70ee768e090d51ee61d515
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>